Rename Composer namespace to SimpleIO since that is all that is left
